Improve coat layer discussion (#19)
Added explanation of what the coat_color means in terms of volumetric absorption.

Also some discussion of the darkening/roughening effects. I opted to remove for now the coat_affect_ parameters. It's difficult to explain how these are supposed to work physically when turned to 0. I see some evidence (in our ADSK internal Slack) that these parameters are actually quite obscure and not asked for or used/understood by artists, but just a legacy of a previous model. So i'd propose we leave them out for now, and gather feedback about whether this causes anyone problems. If it does, we can easily put them back and then at least we will know they are useful, rather than just assuming they might be.