Miscellaneous tweaks made during proofreading (#61)
Several of the changes in this PR are related to the hyphenation of compound adjectives. The general guideline is to hyphenate compound adjectives in situations where there might be ambiguity and not hyphenate where there is no ambiguity:
- hyphenate compound adjectives that come before the nouns they describe
- do not hyphenate compound adjectives that appear at the end of the sentence 
- do not hyphenate compound adjectives in which the first word is an adverb

This PR also updates a couple TODOs, tweaks the lists of example materials, uses American spelling for a couple words, and updates a couple parameter names mentioned in the text.