Remove Standard Surface approximation of coat darkening
This changelist removes the legacy Standard Surface approximation of the darkening and saturation of base colors under coatings.

In a future version of OpenPBR, our plan is to implement this coat darkening in a more subtle, physically based fashion, but the current approximation is too far from this ground truth to be useful.