Update behavior of specular_weight, specular_ior_level, coat_ior_level (#157)
The `specular_ior_level` param is removed, and replaced by `specular_weight` which works in the same way. I incorporated the facility to increase the `specular_weight` to values >1, though with a soft-max at 1 -- retaining the facility we had before, though actually more flexible since it doesn't get capped at an arbitrary maximum value of 2 (it just gets capped by the requirement that the Fresnel factor cannot exceed 1). This seems reasonable and potentially convenient (though we may want to verify that it behaves OK in practice).

The `specular_weight` is now no longer involved in the Fresnel factor multiplier, only `specular_color`. We state that this is applied only to one specific scattering mode (the initial Fresnel reflection from above), and call it out explicitly as an unphysical convenience.