Enable diffuse energy compensation
This changelist adds the `energy_compensation` setting to `oren_nayar_diffuse_bsdf` in the reference implementation for OpenPBR, leveraging the upcoming support for this feature in MaterialX 1.39.

In recent MaterialX TSC meetings, we've aligned on this naming convention for the new energy compensation feature in oren_nayar_diffuse_bsdf, and it now seems safe to move forward with this setting in OpenPBR v1.0.