Diffuse base_weight description changed to match MaterialX EON implementation (#205)
As discussed in https://github.com/AcademySoftwareFoundation/MaterialX/pull/1822#issuecomment-2127738467, the MaterialX implementation assumed that the `base_weight` acts as an overall multiplier of the EON BRDF. While in the OpenPBR spec, we had `base_weight` instead multiplying `base_color`.

In fact the MaterialX intepretation seems better, since then the `base_weight` functions as a linear modulation of the overall albedo, rather than doing something non-linear and dependent on the color. This PR makes the necessary changes to the wording/math of the spec to implement this.